+++ tests/test_dump.py
@@ -1,4 +1,5 @@
 # -*- coding: utf-8 -*-
+
 import contextlib
 import io
 import json